requirements_old.txt
bleach
boto
celery==3.1.23
coverage
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.24.1
django-bootstrap3==6.2.2
django-compressor==2.0
django-contrib-comments==1.6.2
django-countries==3.4.1
django-easy-timezones==0.7.0
django-filter==0.11.0
django-gulp==2.3.0
django-haystack==2.4.1
django-model-utils==2.4
django-notifications-hq==1.2
django-nose
django-postgrespool==0.3.0
django-solo
django-summernote==0.6.16
Django<1.9
elasticsearch==2.3.0
git+https://github.com/jschneier/django-storages.git#egg=django-storages
gunicorn==19.1.1
bpython
newrelic==2.60.0.46
pillow
psycopg2>=2.6
redis==2.10.5
sqlparse==0.1.18
whitenoise==2.0.2
yanc
